Move currency_data and fiat_conversion to their respected folders
