+++ lib/bloc/account/account_model.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:breez/bloc/user_profile/currency.dart';
-import 'package:breez/bloc/user_profile/fiat_conversion.dart';
+import 'package:breez/bloc/account/fiat_conversion.dart';
 import 'package:breez/services/breezlib/data/rpc.pb.dart';
 import 'package:fixnum/fixnum.dart';
