+++ lib/bloc/account/fiat_conversion.dart
@@ -1,4 +1,4 @@
-import 'package:breez/bloc/user_profile/currency_data.dart';
+import 'package:breez/services/currency_data.dart';
 
 class FiatConversion {
   CurrencyData currencyData;
